feat(SharedThematic) revue ajout des couches biodiversite

+++ src/app/enquete/pages/enquete-new/enquete-new.component.ts
@@ -4,7 +4,7 @@ import { map } from 'rxjs';
 
 import { MapContextService } from '../../../shared-map/services/map-context.service';
 import { BreadcrumbTransformerService } from '../../../shared-design-dsfr/transformers/breadcrumb-transformer.service';
-import { INTERSECTED_LAYERS, THEMATICS } from '../../../shared-thematic/models/thematic.enum';
+import { THEMATIC_FICHE_LIST } from '../../../shared-thematic/models/thematic-fiche-list';
 
 @Component({
   selector: 'app-enquete-new',
@@ -49,39 +49,36 @@ export class EnqueteNewComponent implements OnInit {
 
   nextStep() {
     this.step++;
-    switch(this.step) {
-      case 2:
-        if(!this.mapContextService.getLayerDessin().getSource().getFeatures().length) {
-          alert("Veuillez préciser le périmètre de votre forêt à l'aide des outils de dessins disponible sur la carte.");
-          this.step--;
-        }else {
-          this.mapContextService.removeDrawingTools();
-          for(let i = 0; i < THEMATICS.length; i++) {
-            THEMATICS[i].checked = true;
-          }
-          this.mapContextService.updateLayers();
-        };
+    if (this.step === 2) {
+      if (!this.mapContextService.getLayerDessin().getSource().getFeatures().length) {
+        alert("Veuillez préciser le périmètre de votre forêt à l'aide des outils de dessins disponible sur la carte.");
+        this.step--;
         return;
+      }
+      this.mapContextService.removeDrawingTools();
+      for (let i = 0; i < THEMATIC_FICHE_LIST.length; i++) {
+        THEMATIC_FICHE_LIST[i].active = true;
+      }
+      this.mapContextService.updateLayers();
     }
   }
 
 
   previousStep() {
     this.step--;
-    switch(this.step) {
+    switch (this.step) {
       case 0:
         this.mapContextService.resetDessin();
         return;
       case 1:
         this.mapContextService.addDrawingTools();
-        for(let i = INTERSECTED_LAYERS.length; i >= 0; i--) {
-          INTERSECTED_LAYERS.pop();
+        for (let i = this.mapContextService.getActiveThematicLayers().length; i >= 0; i--) {
+          this.mapContextService.getActiveThematicLayers().pop();
         }
-        for(let i = 0; i < THEMATICS.length; i++) {
-          THEMATICS[i].checked = false;
+        for (let i = 0; i < THEMATIC_FICHE_LIST.length; i++) {
+          THEMATIC_FICHE_LIST[i].active = false;
         }
         this.mapContextService.updateLayers();
-        
     }
   }
